server/requirements.txt
eventlet
python-socketio
Werkzeug